Add default security policy
blocks on Lind-Project/lind-wasm-docs#38

Add default security policy file for all projects under the Lind GitHub
organization as per the GitHub community standards. The added policy
file contains a reference to the actual security policy hosted on the
project website plus context info about lind-wasm (see below).

IMPORTANT: The security policy is relevant for the lind-wasm monorepo
only, which is the focus of lind development (see
Lind-Project/lind-wasm#143). I think it makes sense to steer everyone
there, even if they are comig from a different repo in the Lind org.

If that seems undesirable to others, I'm happy to move the file
to the lind-wasm repo.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>